+++ CHANGELOG.md
@@ -4,25 +4,45 @@ Use spec: https://common-changelog.org/
 
 ## Staged
 
+## v0.8.1 - 2026-09-05
+
+### Fixed
+
+- Make ghostty's max_scrollback_lines canonical source of scrollback buffer
+
+## v0.8.0 - 2026-09-03
+
 ### Added
 
 - We now track cwd changes via OSC7
 - Replay window title on attach
 - `ZMX_NO_DETACH_KEY` env var to disable `ctrl+\` keybinding
+- Added `--labels` flag to attach command
+  - e.g. `zmx attach --labels "project=pico env=prod" pico`
+  - e.g. `zmx attach --labels "$(zmx get pico)" pico.sub`
+- New command `zmx print-env` which prints the leader's tracked env vars
 
 ### Fixed
 
 - Clear screen when switching sessions to prevent term state corruption
 - Stray NUL byte in the OSC 7 sequence replayed on attach
 - The OSC 7 cwd is now decoded before the chdir, so a new session can start in a directory whose name needed percent-encoding
+- CLI commands now respond with correct error exist codes
+- Various re-attach handling bugs
 
 ### Changed
 
 - Upgraded to zig v0.16
 - `zmx list` replaced `start_dir` with `cwd`
 - Storing 2k lines of scrollback buffer (like tmux) for each session
-- ZMX_TASK_COMPLETED task marker now includes a 4ch hex id to ensure no collisions from nested task runs
+- `ZMX_TASK_COMPLETED` task marker now includes a 4ch hex id to ensure no collisions from nested task runs
   - `ZMX_TASK_COMPLETED:{id}:0`
+
+## v0.7.1 - 2026-08-27
+
+### Fixed
+
+- Prevent `v0.6.0` daemons from hanging when connecting from `v0.7.0` clients
 
 ## v0.7.0 - 2026-07-23